feat(test): in-source testing for cache
