+++ packages/unplugin-typia/src/core/cache.ts
@@ -1,4 +1,6 @@
 import { accessSync, constants, existsSync, mkdirSync, readFileSync, rmSync, writeFileSync } from 'node:fs';
+import { tmpdir } from 'node:os';
+import process from 'node:process';
 import { createHash } from 'node:crypto';
 import { createRequire } from 'node:module';
 import { basename, dirname, join } from 'pathe';
@@ -126,4 +128,67 @@ function getCacheDir(): FilePath {
 	}
 
 	return wrap<FilePath>(cacheDir);
+}
+
+if (import.meta.vitest != null) {
+	process.env.CACHE_DIR = tmpdir();
+
+	function removeComments(data: string | undefined) {
+		if (data == null) {
+			return data;
+		}
+
+		// eslint-disable-next-line regexp/no-unused-capturing-group
+		return data.replace(/\/\*[\s\S]*?\*\/\n?|([^\\:]|^)\/\/.*$/gm, '');
+	}
+
+	it('return null if cache is not found', async () => {
+		const random = Math.random().toString();
+		const source = wrap<Source>(random);
+		using cache = new Cache(wrap<ID>(random), source);
+		expect(cache.data).toBe(undefined);
+	});
+
+	it('set and get cache', async () => {
+		const random = Math.random().toString();
+		const source = wrap<Source>(random);
+		const filename = wrap<ID>(`${random}-${random}.json`);
+		const data = wrap<Data>(`${random};`);
+
+		/* set cache */
+		{
+			using cache = new Cache(filename, source);
+			cache.data = data;
+		}
+
+		/* get cache */
+		using cache = new Cache(filename, source);
+
+		/* delete js asterisk comments */
+		const cacheDataStr = removeComments(cache.data);
+
+		expect(cacheDataStr).toBe(data);
+	});
+
+	it('set and get null with different id', async () => {
+		const random = Math.random().toString();
+		const source = wrap<Source>(random);
+		const filename = wrap<ID>(`${random}-${random}.json`);
+		const data = wrap<Data>(`${random};`);
+
+		/* set cache */
+		{
+			using cache = new Cache(filename, source);
+			cache.data = data;
+		}
+
+		/* get cache */
+		using cache = new Cache(wrap<ID>(`111;${random}`), source);
+
+		/* delete js asterisk comments */
+		const cacheDataStr = removeComments(cache.data);
+
+		expect(cacheDataStr).toBe(undefined);
+		expect(cacheDataStr).not.toBe(data);
+	});
 }
